restructure manager
