+++ services/manager/internal/service/create_desired_state_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/berops/claudie/internal/manifest"
 	"github.com/berops/claudie/proto/pb/spec"
-	"github.com/berops/claudie/services/manager/server/internal/store"
+	"github.com/berops/claudie/services/manager/internal/store"
 	"github.com/stretchr/testify/assert"
 
 	"google.golang.org/protobuf/proto"
