+++ services/manager/internal/service/grpc.go
@@ -12,7 +12,7 @@ import (
 	"github.com/berops/claudie/internal/envs"
 	"github.com/berops/claudie/internal/utils"
 	"github.com/berops/claudie/proto/pb"
-	"github.com/berops/claudie/services/manager/server/internal/store"
+	"github.com/berops/claudie/services/manager/internal/store"
 	"github.com/rs/zerolog/log"
 )
