+++ services/manager/internal/service/handler_task_update.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/berops/claudie/proto/pb"
 	"github.com/berops/claudie/proto/pb/spec"
-	"github.com/berops/claudie/services/manager/server/internal/store"
+	"github.com/berops/claudie/services/manager/internal/store"
 	"github.com/rs/zerolog/log"
 
 	"google.golang.org/grpc/codes"
