+++ services/manager/internal/service/handler_update_nodepool.go
@@ -11,7 +11,7 @@ import (
 	"github.com/berops/claudie/internal/utils"
 	"github.com/berops/claudie/proto/pb"
 	"github.com/berops/claudie/proto/pb/spec"
-	"github.com/berops/claudie/services/manager/server/internal/store"
+	"github.com/berops/claudie/services/manager/internal/store"
 	"github.com/google/uuid"
 	"github.com/rs/zerolog/log"
