+++ services/manager/internal/service/handler_upsert_manifest.go
@@ -14,7 +14,7 @@ import (
 	"github.com/berops/claudie/internal/checksum"
 	"github.com/berops/claudie/internal/manifest"
 	"github.com/berops/claudie/proto/pb"
-	"github.com/berops/claudie/services/manager/server/internal/store"
+	"github.com/berops/claudie/services/manager/internal/store"
 	"github.com/rs/zerolog/log"
 )
