+++ services/manager/internal/service/schedule_tasks.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/berops/claudie/internal/utils"
 	"github.com/berops/claudie/proto/pb/spec"
-	"github.com/berops/claudie/services/manager/server/internal/store"
+	"github.com/berops/claudie/services/manager/internal/store"
 	"github.com/google/uuid"
 	"github.com/rs/zerolog/log"
