+++ services/manager/internal/service/watchers.go
@@ -8,7 +8,7 @@ import (
 	"github.com/berops/claudie/internal/manifest"
 	"github.com/berops/claudie/internal/utils"
 	"github.com/berops/claudie/proto/pb/spec"
-	"github.com/berops/claudie/services/manager/server/internal/store"
+	"github.com/berops/claudie/services/manager/internal/store"
 )
 
 // TaskTTL is the minimum number of ticks (every ~10sec) within which a given task must be completed
