+++ services/manager/internal/service/watchers_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/berops/claudie/internal/checksum"
 	"github.com/berops/claudie/internal/manifest"
 	"github.com/berops/claudie/proto/pb/spec"
-	"github.com/berops/claudie/services/manager/server/internal/store"
+	"github.com/berops/claudie/services/manager/internal/store"
 	"github.com/stretchr/testify/assert"
 )
