+++ services/manager/internal/store/in_memory_store.go
